pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guideai"
version = "0.1.0"
description = "GuideAI developer tooling and task orchestration helpers"
readme = "docs/README.md"
requires-python = ">=3.10"
keywords = ["guideai", "cli", "developer-tools"]
license = { text = "Proprietary" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Environment :: Console",
  "Intended Audience :: Developers",
]
authors = [
  { name = "GuideAI Team" }
]
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn>=0.30,<1.0",
  "httpx>=0.27,<0.29",
  "click>=8.1,<9.0",
  "keyring>=24,<25",
  "python-multipart>=0.0.6,<1.0",
  "pydantic>=2.5,<3.0",
  "pydantic-settings>=2.1,<3.0",
  "boto3>=1.28,<2.0",  # AWS SDK for S3/MinIO storage
  "PyJWT>=2.8,<3.0",  # JWT tokens for internal auth
  "bcrypt>=4.1,<5.0",  # Password hashing
]

[project.urls]
Documentation = "https://example.com/guideai"
Source = "https://github.com/Nas4146/guideai"

[project.optional-dependencies]
dev = [
  "pytest>=7.4,<9.0",
  "pytest-cov>=4.1,<6.0",
  "pytest-asyncio>=0.21,<1.0",
  "pytest-xdist>=3.5,<4.0",  # Parallel test execution
  "pytest-timeout>=2.2,<3.0",  # Test timeouts
  "pytest-mock>=3.12,<4.0",  # Mocking utilities
  "black>=23.0,<25.0",
  "isort>=5.12,<6.0",
  "flake8>=6.0,<8.0",
  "mypy>=1.5,<2.0",
]
telemetry = [
  "kafka-python>=2.0,<3.0",
  "duckdb>=1.1,<2.0",
  "pytz>=2023.3",
  "psycopg2-binary>=2.9,<3.0",
]
redis = [
  "redis>=5.0,<6.0",
]
postgres = [
  "sqlalchemy>=2.0,<3.0",
  "psycopg2-binary>=2.9,<3.0",
  "prometheus_client>=0.19,<1.0",
]
semantic = [
  "sentence-transformers>=2.2.0,<3.0",
  "faiss-cpu>=1.7.0,<2.0",
]
amprealize = [
  "amprealize>=0.1.0",
]

[project.scripts]
guideai = "guideai.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
minversion = "7.4"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=guideai",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=xml",
    "--cov-report=html",
    "--cov-fail-under=70",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "parity: marks tests as service parity tests",
    "asyncio: marks tests as async",
]
asyncio_mode = "auto"
timeout = 30
# Parallel execution: use -n 2 or -n 4 (safer than -n auto on laptops)
# Recommended: pytest -n 2 for laptops with limited resources
# CI can use: pytest -n auto

[tool.coverage.run]
source = ["guideai"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
